renamed stencils graph classes for consitency
