+++ src/hhimg/algorithm/tile/update_stencils_task.h
